# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Monitoring API for Decentralized Sequencer Validator pipeline",
        "title": "DSV Pipeline Monitor API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/aggregation/queue": {
            "get": {
                "description": "Get aggregation queue status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aggregation"
                ],
                "summary": "Aggregation queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol state identifier",
                        "name": "protocol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data market address",
                        "name": "market",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/aggregation/results": {
            "get": {
                "description": "Get network-wide aggregated batches with complete batch information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aggregation"
                ],
                "summary": "Aggregated batches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol state identifier",
                        "name": "protocol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data market address",
                        "name": "market",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.AggregatedBatch"
                            }
                        }
                    }
                }
            }
        },
        "/batches/finalization-queue": {
            "get": {
                "description": "Get finalization queue status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "Finalization queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol state identifier",
                        "name": "protocol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data market address",
                        "name": "market",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/batches/finalized": {
            "get": {
                "description": "Get finalized batches with details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "Finalized batches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol state identifier",
                        "name": "protocol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data market address",
                        "name": "market",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.FinalizedBatch"
                            }
                        }
                    }
                }
            }
        },
        "/batches/parts": {
            "get": {
                "description": "Get completed batch parts from parallel finalization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "Batch parts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol state identifier",
                        "name": "protocol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data market address",
                        "name": "market",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.BatchPart"
                            }
                        }
                    }
                }
            }
        },
...
        "/pipeline/overview": {
            "get": {
                "description": "Get overall pipeline status summary",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Pipeline overview",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol state identifier",
                        "name": "protocol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data market address",
                        "name": "market",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PipelineOverview"
                        }
                    }
                }
            }
        },
        "/submissions/queue": {
            "get": {
                "description": "Get submission queue depth and sample entries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submissions"
                ],
                "summary": "Submission queue status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol state identifier",
                        "name": "protocol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data market address",
                        "name": "market",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/workers/status": {
            "get": {
                "description": "Get status of all finalization workers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Worker status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol state identifier",
                        "name": "protocol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data market address",
                        "name": "market",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.WorkerStatus"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.AggregatedBatch": {
            "type": "object",
            "properties": {
                "batch_ipfs_cid": {
                    "description": "IPFS CID of the final Level 2 aggregated batch",
                    "type": "string"
                },
                "epoch_id": {
                    "type": "string"
                },
                "merkle_root": {
                    "description": "Merkle root as base64 string",
                    "type": "string"
                },
                "project_count": {
                    "type": "integer"
                },
                "project_ids": {
                    "description": "Array of project IDs that were aggregated",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "timestamp": {
                    "type": "string"
                },
                "validator_batches": {
                    "description": "validator_id → their individual Level 1 batch IPFS CID",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "validator_count": {
                    "type": "integer"
                },
                "validator_ids": {
                    "description": "Array of validator IDs that contributed",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "main.BatchPart": {
            "type": "object",
            "properties": {
                "epoch_id": {
                    "type": "string"
                },
                "part_id": {
                    "type": "integer"
                },
                "project_count": {
                    "type": "integer"
                }
            }
        },
        "main.FinalizedBatch": {
            "type": "object",
            "properties": {
                "epoch_id": {
                    "type": "string"
                },
                "ipfs_cid": {
                    "type": "string"
                },
                "project_count": {
                    "type": "integer"
                },
                "sequencer_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "vote_count": {
                    "type": "integer"
                }
            }
        },
        "main.PipelineOverview": {
            "type": "object",
            "properties": {
                "active_windows": {
                    "type": "integer"
                },
                "active_workers": {
                    "type": "integer"
                },
                "aggregated_batches": {
                    "type": "integer"
                },
                "aggregation_queue": {
                    "$ref": "#/definitions/main.QueueStatus"
                },
                "completed_parts": {
                    "type": "integer"
                },
                "finalization_queue": {
                    "$ref": "#/definitions/main.QueueStatus"
                },
                "finalized_batches": {
                    "type": "integer"
                },
                "ready_batches": {
                    "type": "integer"
                },
                "submission_queue": {
                    "$ref": "#/definitions/main.QueueStatus"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "main.WorkerStatus": {
            "type": "object",
            "properties": {
                "epoch_id": {
                    "type": "string"
                },
                "healthy": {
                    "type": "boolean"
                },
                "last_heartbeat": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        }
    }
}